pyproject.toml
[tool.poetry]
name = "academyconnect"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.1"
drf-yasg = "^1.21.7"
coreapi = "^2.3.3"
django-cors-headers = "^4.2.0"
stripe = "^6.5.0"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
kombu = "^5.3.2"
redis = "^5.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
